Fix issues with build toolchain.
builder:
Fix bug in toolchain selection when system has a toolchain installed that is postfixed with a version number.

other:
Added libzstd link library to CGO_LDFLAGS that is required on some systems.
Added toolchain file that contains a bypass for a shared library error raised by CMake when compiling compiler-rt.
Updated Makefile to pass the toolchain file to the CMake command for configuring compiler-rt.
Updated README.