Fixed bug in heap escape analysis.
goir:
Fixed an issue with the heap escape analysis where it would move a stack allocation to the heap because it was involved in a method call that has a pointer receiver.
Improved heap escape analysis to detect if the usage of a receiver value in method should cause the respective stack allocation in the caller to escape to the heap.