Refactor mutex to use a ticketing mechanism to reduce lock contention.
The previous implementation favored faster goroutines which allow slower goroutines to be starved by lock contention. The new implementation uses a simple ticketing mechanism to guarantee that another goroutine waiting on the lock is allowed to acquire the lock when the lock is released.

goir:
Refactored `AtomicAddIOp` lowering to yield the new value (Go's expectation) instead of the old one (LLVM's behavior).

sync:
Implemented ticketing mechanism in mutex.
Added double-lock and double-free panics to mutex implementation.
Added panic when mutex is attempted to be unlocked by a goroutine other than the one that originally acquired its lock.